Bytecode VM: Move BytecodeRoData into a subdirectory
