+++ lib/natalie/compiler.rb
@@ -3,7 +3,7 @@ require_relative '../../build/generated/numbers'
 require_relative './compiler/backends/cpp_backend'
 require_relative './compiler/bytecode'
 require_relative './compiler/bytecode/loader'
-require_relative './compiler/bytecode_ro_data'
+require_relative './compiler/bytecode/ro_data'
 require_relative './compiler/comptime_values'
 require_relative './compiler/instruction_manager'
 require_relative './compiler/loaded_file'
@@ -54,7 +54,7 @@ module Natalie
     end
 
     def compile_to_bytecode(io)
-      rodata = BytecodeRoData.new
+      rodata = Bytecode::RoData.new
 
       bytecode = instructions.each.with_object(''.b) do |instruction, output|
         output << instruction.serialize(rodata)
